Remove generated c file and register node addon in root.zig
Node will look for an exported symbol named: napi_register_module_v1

This symbol is a function with the same signature as the init function.
In C the standard is to use the NAPI_MODULE macro. This ultimately gets
expanded to what we are exporting.

If zig had better support for using c macros then I would most likely
choose that instead. Howevew since this is still very experimental I
like being able to have everything be zig.